Split each input format into their respectie file
