+++ react/FormInput.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { InputProps } from './components/InputTypes'
-import { FieldBuilder } from './components/Inputs'
+import { FieldBuilder } from './components/FieldBuilder'
 
 export default function FormInput(props: InputProps): React.ReactNode {
   return <FieldBuilder {...props} />
