+++ react/modules/JSONPathHandler.ts
@@ -4,12 +4,22 @@ const concatFormPath = (path: string, newNode: string): string => {
   return path + '/' + newNode
 }
 
+const getSplitPath = (path: string): Array<string> => {
+  const split = path.split('/')
+  if (split[0] === '#') {
+    split.shift()
+  }
+  if (split[split.length - 1] === '') {
+    split.pop()
+  }
+  return split
+}
+
 const getObjectFromPath = (
   obj: JSONSchemaType,
   path: string
 ): JSONSchemaType | undefined => {
-  const splitPath = path.split('/')
-  splitPath.shift()
+  const splitPath = getSplitPath(path)
   let currentOriginal: JSONSchemaType | undefined = obj
 
   for (let node = 0; node < splitPath.length; node++) {
@@ -34,8 +44,7 @@ const buildObjectFromFormData = (
   const objectFromData: JSONSchemaType = {}
 
   for (const key of orderedSchemaKeys) {
-    const splitPath = key.split('/')
-    splitPath.shift()
+    const splitPath = getSplitPath(key)
 
     let current = objectFromData
     let currentOriginal = originalSchema
@@ -86,4 +95,27 @@ const buildObjectFromFormData = (
   return objectFromData
 }
 
-export { buildObjectFromFormData, concatFormPath, getObjectFromPath }
+const isRequiredField = (obj: JSONSchemaType, path: string): boolean => {
+  let required = false
+  const splitPath = getSplitPath(path)
+  const childName = splitPath.pop()
+  if (childName) {
+    const parentPath = '#/' + splitPath.join('/')
+    const parentObject = getObjectFromPath(obj, parentPath)
+    if (
+      parentObject &&
+      parentObject.required &&
+      (parentObject.required as Array<string>).indexOf(childName) > -1
+    ) {
+      required = true
+    }
+  }
+  return required
+}
+
+export {
+  buildObjectFromFormData,
+  concatFormPath,
+  getObjectFromPath,
+  isRequiredField,
+}
